feat: created .protobuf file
