--- calc.go
@@ -1,4 +1,0 @@
-package calc // import "ultre.me/calcbiz"
-
-// VERSION represents the version of the Camembert au lait crew's website
-const VERSION = "2.1.0"
