+++ go.mod
@@ -2,7 +2,7 @@ module ultre.me/calcbiz
 
 require (
 	github.com/gin-contrib/sse v0.0.0-20170109093832-22d885f9ecc7 // indirect
-	github.com/gin-gonic/gin v1.3.0
+	github.com/gin-gonic/gin v1.3.0 // indirect
 	github.com/go-chi/chi v3.3.3+incompatible
 	github.com/go-chi/render v1.0.1
 	github.com/gobuffalo/packr v1.13.7
@@ -18,7 +18,7 @@ require (
 	github.com/shazow/memoizer v0.0.0-20130904030615-74fc48eaeadc
 	github.com/smartystreets/assertions v0.0.0-20180927180507-b2de0cb4f26d // indirect
 	github.com/smartystreets/goconvey v0.0.0-20180222194500-ef6db91d284a // indirect
-	github.com/tpyolang/tpyo-cli v0.0.0-20160926074508-6ba04c3a8dec
+	github.com/tpyolang/tpyo-cli v0.0.0-20160926074508-6ba04c3a8dec // indirect
 	github.com/ugorji/go/codec v0.0.0-20181022190402-e5e69e061d4f // indirect
 	github.com/urfave/cli v1.20.0
 	github.com/yanatan16/golang-soundcloud v0.0.0-20161013200625-a182dd459b66
@@ -26,7 +26,7 @@ require (
 	golang.org/x/sys v0.0.0-20181023152157-44b849a8bc13 // indirect
 	gopkg.in/go-playground/assert.v1 v1.2.1 // indirect
 	gopkg.in/go-playground/validator.v8 v8.18.2 // indirect
-	ultre.me/kryptos v0.0.0-20181023194748-240fe1a16033
-	ultre.me/moi-j-aime-generator v0.0.0-20181023194533-7b56f5b329e1
-	ultre.me/recettator v0.4.0
+	ultre.me/kryptos v0.0.0-20181023194748-240fe1a16033 // indirect
+	ultre.me/moi-j-aime-generator v0.0.0-20181023194533-7b56f5b329e1 // indirect
+	ultre.me/recettator v0.4.0 // indirect
 )
